build: add http2 reverse proxy
