+++ server/http2-proxy/server.crt
@@ -1,0 +1,21 @@
+-----BEGIN CERTIFICATE-----
+MIIDazCCAlOgAwIBAgIUGQmgc5Q6gq3jDvWaSNccaIeVHIIwDQYJKoZIhvcNAQEL
+BQAwRTELMAkGA1UEBhMCRlIxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
+GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDAeFw0xOTAzMTAxODA5MTlaFw0xOTA0
+MDkxODA5MTlaMEUxCzAJBgNVBAYTAkZSMRMwEQYDVQQIDApTb21lLVN0YXRlMSEw
+HwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQC8ufdm40a5W2rbjJtJpjUEgyqja/0FYhFwEiQDkGnI
+uAMKJd51xyvkUXWTcxe0WyS8dS2406gqdULvWoze8KEGcjsd0YCF74nGS0Tiv8wR
+IJlvOmAB8+/firqD1xOQ2ySIp2w7OtBpbYOwYVgcvCTcNkmVTsBofUXwGjpQLOus
+j/7CwKrvCe8qAtIXNy/QhMZAnT/xbNdB12+fR+T1MimSgy0Mku+RVAt3oWECLJSY
+vEO4nqlO0FnRNpEJnWJfX4CTzrv84LAfod+6sc5Ze2OuP6RTbgfPXobXcE0EnZH0
+fmMHhHZ0c13DB0sQHKMcnCjo3vBzZgDBz2EktQRNaLofAgMBAAGjUzBRMB0GA1Ud
+DgQWBBQ+LCl4rDwrPaSwLuuwdYKgZ8w2szAfBgNVHSMEGDAWgBQ+LCl4rDwrPaSw
+LuuwdYKgZ8w2szAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQAL
+VHFfWTxntwfn4vj1VKSII7V/WdqnZgGot0Sqzp6iQ1PQFDfpB4X726CsUhatKViY
+JNyCm8uwj/FkYQRrj9GyHq5quMma+CCWsT6QB0SLwhiXolxDyM3sALj+m/T2rlW3
+bVy/YBvT5v/0gR+YQ59HooIzr6vsgA9al7S/uJSWCHT97Gwee1qoioQoJSfKX0PS
+E7zTm5Xp404AXJHuD6vZ5FpmGYU6dyF64BswKlxV/KbzMfatK8SgGFua2psFIhjs
+2PAASytH3yG5ISFfeAeLGDm2rv+lW65rs1agdm9mZKX1ZqIz2QYpyRABJ4jZJBa9
+NJo0H5wD3fwgFgYgRPc4
+-----END CERTIFICATE-----
